refactor: extract fetchLatestCliVersion into shared lib/npm.ts
